Add s:SlimeSendRaw to plugin/slime.vim
I asked about this at the end of last year, and there was a commit
pushed against it, though the issue remains open, and the fix doesn't
work for me. I threw this hack in there to let me send a raw command,
regardless of filetype, similar to my patch for the Vimya plugin.